+++ frontend/src/components/messages/Chats.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect, useRef } from "react";
 import styles from "./Chats.module.css";
 import ChatLeft from "./ChatLeft";
 import ChatRight from "./ChatRight";
+import SendMessageForm from "./SendMessageForm";
 
 // 나의 정보가 있어야 한다.
 import myImage from "../../assets/images/profile.png";
@@ -68,40 +69,8 @@ export default function Chats({ id, name, image }) {
         className={`${styles["flex-grow-0"]} ${styles["py-3"]} ${styles["px-4"]} ${styles["border-top"]}`}
       >
         <SendMessageForm onSendMessage={handleSendMessage} />
-        {/* <div className="input-group">
-          <input
-            type="text"
-            className="form-control m-"
-            placeholder="메세지를 입력해주세요."
-          />
-          <button className="btn btn-primary">전송</button>
-        </div> */}
       </div>
     </React.Fragment>
   );
 }
 
-const SendMessageForm = ({ onSendMessage }) => {
-  const [message, setMessage] = useState("");
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    onSendMessage({ id: myId, text: message, date: new Date() });
-    setMessage("");
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <div className="input-group">
-        <textarea
-          className="form-control"
-          value={message}
-          onChange={(event) => setMessage(event.target.value)}
-        ></textarea>
-        <button type="submit" className="btn btn-primary">
-          전송
-        </button>
-      </div>
-    </form>
-  );
-};
